update : move PrivateChat component
