+++ frontend/src/assets/styles/Chat/private_chat.scss
