+++ frontend/src/assets/styles/Room/room_list.scss
@@ -7,12 +7,6 @@
         box-shadow: rgba(0, 0, 0, 0.08) 0px 4px 12px;
     }
 
-    .create-btn-text {
-        margin-left: 20px;
-        font-weight: bold;
-        font-size: 17px;
-    }
-
     .plus-btn {
         margin-left: 10px;
     }
