+++ frontend/src/assets/styles/main.scss
@@ -12,6 +12,6 @@
 @import "./Register/register";
 @import "./Loading/loading";
 @import "./Chat/group_chat";
+@import "./Chat/private_chat";
 @import "./Error/error";
 @import "./Error/empty";
-@import "./PrivateChat/private_chat";
